// ec/src/models/api/msm/ect/series/SearchSeriesResponse$search.ts
export const SpecViewType = {
	TEXT_BUTTON: '1',
	TEXT_SINGLE_LINE: '2',
	TEXT_DOUBLE_LINE: '3',
	TEXT_TRIPLE_LINE: '4',
	IMAGE_SINGLE_LINE: '5',
	IMAGE_DOUBLE_LINE: '6',
	IMAGE_TRIPLE_LINE: '7',
	NUMERIC: '8',
	LIST: '9',
	TREE: '10',
	AGGREGATION: '11',
	CENTER: '12',
} as const;
export type SpecViewType = typeof SpecViewType[keyof typeof SpecViewType];